Correcciones esteticas

+++ src/components/Administracion.jsx
@@ -9,7 +9,7 @@ import { usuariosGet, usuarioDelete } from "../helpers/usuarios";
 import ModalUsuarios from "./ModalUsuarios";
 import ModalProductos from "./ModalProductos";
 import ModalComandas from "./ModalComanda"
-
+import Swal from "sweetalert2";
 const Administracion = () => {
   const [render, setRender] = useState(false);
   const [toggleProducto, setToggleProducto] = useState(false);
@@ -27,12 +27,30 @@ const Administracion = () => {
   const history = useHistory();
 
   const handleDeleteProducto = (product) => {
-    deleteProducto(product._id, token).then((respuesta) => {
-      if (respuesta.msg) {
-        window.alert(respuesta.msg);
-        setRender(!render);
-      }
-    });
+    Swal.fire({
+      title: "¿Seguro que quieres eliminar el producto?",
+      text: "No puedes revertir esta acción",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes",
+      cancelButtonText: "No",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        deleteProducto(product._id, token).then((respuesta) => {
+          if (respuesta.msg) {
+            Swal.fire(
+              {
+                title: respuesta.msg,
+                text: "Operacion exitosa",
+                icon: "success",
+                confirmButtonColor: "#3085d6",
+              });
+            setRender(!render);
+          }
+        });
+      }})
   };
   const handleEditProducto = (product) => {
     setProductEditar(product);
@@ -229,7 +247,7 @@ const Administracion = () => {
   ];
 
   return (
-    <div className="bg">
+    <div className="bg mt-5 pt-3">
       <div className="d-flex align-items-center">
         <h5 className="text-white p-4">PRODUCTOS</h5>
         <button
